+++ lib/mediawiki/butt.rb
@@ -9,20 +9,12 @@ module MediaWiki
     include MediaWiki::Query::Meta
     include MediaWiki::Query::Properties
     include MediaWiki::Query::Lists
-    # Creates a new instance of MediaWiki::Butt
+
+    # Creates a new instance of MediaWiki::Butt. To work with any MediaWiki::Butt methods, you must first create an instance of it.
     #
-    # ==== Attributes
-    #
-    # * +url+ - The FULL wiki URL. api.php can be omitted, but it will make harsh assumptions about your wiki configuration.
-    # * +use_ssl+ - Whether or not to use SSL. Will default to true.
-    #
-    # ==== Examples
-    #
-    # The example below shows an ideal usage of the method.
-    # => butt = MediaWiki::Butt.new("http://ftb.gamepedia.com/api.php")
-    #
-    # The example below shows a less than idea, but still functional, usage of the method. It is less than ideal because it has to assume that your API page is at /api.php, but it could easily be at /w/api.php, or even /wiki/api.php. It also does not use a secure connection.
-    # => butt = MediaWiki::Butt.new("http://ftb.gamepedia.com", false)
+    # @param url [String] The FULL wiki URL. api.php can be omitted, but it will make harsh assumptions about your wiki configuration.
+    # @param use_ssl [Boolean] Whether or not to use SSL. Will default to true.
+    # @return [MediaWiki::Butt] new instance of MediaWiki::Butt
     def initialize(url, use_ssl = true)
       if url =~ /api.php$/
         @url = url
@@ -39,15 +31,10 @@ module MediaWiki
 
     # Performs a generic HTTP POST action and provides the response. This method generally should not be used by the user, unless there is not a method provided by the Butt developers for a particular action.
     #
-    # ==== Attributes
-    #
-    # * +params+ - A basic hash containing MediaWiki API parameters. Please see mediawiki.org/wiki/API for more information.
-    # * +autoparse+ - Whether or not to provide a parsed version of the response's JSON. Will default to true.
-    # * +setcookie+ - Whether you want to set the auth cookie. Only used in authentication. Defaults to false.
-    #
-    # ==== Examples
-    #
-    # => login = butt.post({action: 'login', lgname: username, lgpassword: password, format: 'json'})
+    # @param params [Hash] A basic hash containing MediaWiki API parameters. Please see the MediaWiki API for more information.
+    # @param autoparse [Boolean] Whether or not to provide a parsed version of the response's JSON. Will default to true.
+    # @param setcookie [Boolean] Whether you want to set the auth cookie. Only used in authentication. Defaults to false.
+    # @return [JSON/Response] Parsed JSON if autoparse is true, or raw response if not.
     def post(params, autoparse = true, setcookie = false)
       if setcookie == true
         header = { 'Set-Cookie' => @cookie }
@@ -63,8 +50,8 @@ module MediaWiki
       end
     end
 
-    # Returns true if the currently logged in user is in the "bot" group.
-    # Will return false if the user is either not a bot or if they are not logged in.
+    # Returns true if the currently logged in user is in the "bot" group. This can be helpful to some developers, but it is mostly for use internally in MediaWiki::Butt.
+    # @return [Boolean] true if logged in as a bot, false if not logged in or logged in as a non-bot
     def is_current_user_bot()
       if @logged_in == true
         params = {